# Optimise `import qiskit` with lazy imports