--- a/qiskit/algorithms/optimizers/imfil.py
+++ b/qiskit/algorithms/optimizers/imfil.py
@@ -18,6 +18,7 @@ from qiskit.utils import optionals as _optionals
 from .optimizer import Optimizer, OptimizerSupportLevel, OptimizerResult, POINT
 
 
+@_optionals.HAS_SKQUANT.require_in_instance
 class IMFIL(Optimizer):
     """IMplicit FILtering algorithm.
 
@@ -42,7 +43,6 @@ class IMFIL(Optimizer):
         Raises:
             MissingOptionalLibraryError: scikit-quant not installed
         """
-        _optionals.HAS_SKQUANT.require_now("IMFIL")
         super().__init__()
         self._maxiter = maxiter