--- a/qiskit/algorithms/optimizers/snobfit.py
+++ b/qiskit/algorithms/optimizers/snobfit.py
@@ -15,23 +15,8 @@
 from typing import Any, Dict, Optional, Callable, Tuple, List
 
 import numpy as np
-from qiskit.exceptions import MissingOptionalLibraryError
+from qiskit.utils import optionals as _optionals
 from .optimizer import Optimizer, OptimizerSupportLevel, OptimizerResult, POINT
-
-
-try:
-    import skquant.opt as skq
-
-    _HAS_SKQUANT = True
-except ImportError:
-    _HAS_SKQUANT = False
-
-try:
-    from SQSnobFit import optset
-
-    _HAS_SKSNOBFIT = True
-except ImportError:
-    _HAS_SKSNOBFIT = False
 
 
 class SNOBFIT(Optimizer):
@@ -64,14 +49,8 @@ class SNOBFIT(Optimizer):
         Raises:
             MissingOptionalLibraryError: scikit-quant or SQSnobFit not installed
         """
-        if not _HAS_SKQUANT:
-            raise MissingOptionalLibraryError(
-                libname="scikit-quant", name="SNOBFIT", pip_install="pip install scikit-quant"
-            )
-        if not _HAS_SKSNOBFIT:
-            raise MissingOptionalLibraryError(
-                libname="SQSnobFit", name="SNOBFIT", pip_install="pip install SQSnobFit"
-            )
+        _optionals.HAS_SKQUANT.require_now("SNOBFIT")
+        _optionals.HAS_SQSNOBFIT.require_now("SNOBFIT")
         super().__init__()
         self._maxiter = maxiter
         self._maxfail = maxfail
@@ -102,6 +81,9 @@ class SNOBFIT(Optimizer):
         jac: Optional[Callable[[POINT], POINT]] = None,
         bounds: Optional[List[Tuple[float, float]]] = None,
     ) -> OptimizerResult:
+        import skquant.opt as skq
+        from SQSnobFit import optset
+
         snobfit_settings = {
             "maxmp": self._maxmp,
             "maxfail": self._maxfail,