--- a/qiskit/circuit/parameterexpression.py
+++ b/qiskit/circuit/parameterexpression.py
@@ -20,13 +20,7 @@ import operator
 import numpy
 
 from qiskit.circuit.exceptions import CircuitError
-
-try:
-    import symengine
-
-    HAS_SYMENGINE = True
-except ImportError:
-    HAS_SYMENGINE = False
+from qiskit.utils import optionals as _optionals
 
 
 # This type is redefined at the bottom to insert the full reference to "ParameterExpression", so it
@@ -71,7 +65,9 @@ class ParameterExpression:
 
     def conjugate(self) -> "ParameterExpression":
         """Return the conjugate."""
-        if HAS_SYMENGINE:
+        if _optionals.HAS_SYMENGINE:
+            import symengine
+
             conjugated = ParameterExpression(
                 self._parameter_symbols, symengine.conjugate(self._symbol_expr)
             )
@@ -170,7 +166,9 @@ class ParameterExpression:
 
         self._raise_if_passed_unknown_parameters(parameter_map.keys())
         self._raise_if_parameter_names_conflict(inbound_names, parameter_map.keys())
-        if HAS_SYMENGINE:
+        if _optionals.HAS_SYMENGINE:
+            import symengine
+
             new_parameter_symbols = {p: symengine.Symbol(p.name) for p in inbound_parameters}
         else:
             from sympy import Symbol
@@ -292,7 +290,9 @@ class ParameterExpression:
 
         # Compute the gradient of the parameter expression w.r.t. param
         key = self._parameter_symbols[param]
-        if HAS_SYMENGINE:
+        if _optionals.HAS_SYMENGINE:
+            import symengine
+
             expr_grad = symengine.Derivative(self._symbol_expr, key)
         else:
             # TODO enable nth derivative
@@ -351,7 +351,9 @@ class ParameterExpression:
 
     def sin(self):
         """Sine of a ParameterExpression"""
-        if HAS_SYMENGINE:
+        if _optionals.HAS_SYMENGINE:
+            import symengine
+
             return self._call(symengine.sin)
         else:
             from sympy import sin as _sin
@@ -360,7 +362,9 @@ class ParameterExpression:
 
     def cos(self):
         """Cosine of a ParameterExpression"""
-        if HAS_SYMENGINE:
+        if _optionals.HAS_SYMENGINE:
+            import symengine
+
             return self._call(symengine.cos)
         else:
             from sympy import cos as _cos
@@ -369,7 +373,9 @@ class ParameterExpression:
 
     def tan(self):
         """Tangent of a ParameterExpression"""
-        if HAS_SYMENGINE:
+        if _optionals.HAS_SYMENGINE:
+            import symengine
+
             return self._call(symengine.tan)
         else:
             from sympy import tan as _tan
@@ -378,7 +384,9 @@ class ParameterExpression:
 
     def arcsin(self):
         """Arcsin of a ParameterExpression"""
-        if HAS_SYMENGINE:
+        if _optionals.HAS_SYMENGINE:
+            import symengine
+
             return self._call(symengine.asin)
         else:
             from sympy import asin as _asin
@@ -387,7 +395,9 @@ class ParameterExpression:
 
     def arccos(self):
         """Arccos of a ParameterExpression"""
-        if HAS_SYMENGINE:
+        if _optionals.HAS_SYMENGINE:
+            import symengine
+
             return self._call(symengine.acos)
         else:
             from sympy import acos as _acos
@@ -396,7 +406,9 @@ class ParameterExpression:
 
     def arctan(self):
         """Arctan of a ParameterExpression"""
-        if HAS_SYMENGINE:
+        if _optionals.HAS_SYMENGINE:
+            import symengine
+
             return self._call(symengine.atan)
         else:
             from sympy import atan as _atan
@@ -405,7 +417,9 @@ class ParameterExpression:
 
     def exp(self):
         """Exponential of a ParameterExpression"""
-        if HAS_SYMENGINE:
+        if _optionals.HAS_SYMENGINE:
+            import symengine
+
             return self._call(symengine.exp)
         else:
             from sympy import exp as _exp
@@ -414,7 +428,9 @@ class ParameterExpression:
 
     def log(self):
         """Logarithm of a ParameterExpression"""
-        if HAS_SYMENGINE:
+        if _optionals.HAS_SYMENGINE:
+            import symengine
+
             return self._call(symengine.log)
         else:
             from sympy import log as _log
@@ -480,7 +496,7 @@ class ParameterExpression:
         if isinstance(other, ParameterExpression):
             if self.parameters != other.parameters:
                 return False
-            if HAS_SYMENGINE:
+            if _optionals.HAS_SYMENGINE:
                 from sympy import sympify
 
                 return sympify(self._symbol_expr).equals(sympify(other._symbol_expr))
@@ -491,7 +507,7 @@ class ParameterExpression:
         return False
 
     def __getstate__(self):
-        if HAS_SYMENGINE:
+        if _optionals.HAS_SYMENGINE:
             from sympy import sympify
 
             symbols = {k: sympify(v) for k, v in self._parameter_symbols.items()}
@@ -507,6 +523,8 @@ class ParameterExpression:
 
     def __setstate__(self, state):
         if state["type"] == "symengine":
+            import symengine
+
             self._symbol_expr = symengine.sympify(state["expr"])
             self._parameter_symbols = {k: symengine.sympify(v) for k, v in state["symbols"].items()}
             self._parameters = set(self._parameter_symbols)
@@ -525,7 +543,7 @@ class ParameterExpression:
             # but the parameter will evaluate as real. Check that if the
             # expression's is_real attribute returns false that we have a
             # non-zero imaginary
-            if HAS_SYMENGINE:
+            if _optionals.HAS_SYMENGINE:
                 if self._symbol_expr.imag != 0.0:
                     return False
             else: