--- a/qiskit/opflow/gradients/hessian.py
+++ b/qiskit/opflow/gradients/hessian.py
@@ -17,7 +17,6 @@ import functools
 import numpy as np
 
 from qiskit.circuit.quantumcircuit import _compare_parameters
-from qiskit.exceptions import MissingOptionalLibraryError
 from qiskit.circuit import ParameterVector, ParameterExpression
 from qiskit.utils import optionals as _optionals
 from ..operator_globals import Zero, One
@@ -236,42 +235,20 @@ class Hessian(HessianBase):
                 ]
             )
 
+            _optionals.HAS_JAX.require_now("automatic differentiation")
+            from jax import grad, jit  # pylint: disable=import-error
+
             if operator.grad_combo_fn:
                 first_partial_combo_fn = operator.grad_combo_fn
-                if _optionals.HAS_JAX:
-                    from jax import grad, jit
 
-                    second_partial_combo_fn = jit(
-                        grad(lambda x: first_partial_combo_fn(x)[0], holomorphic=True)
-                    )
-                else:
-                    raise MissingOptionalLibraryError(
-                        libname="jax",
-                        name="get_hessian",
-                        msg=(
-                            "This automatic differentiation function is based on JAX. Please "
-                            "install jax and use `import jax.numpy as jnp` instead of "
-                            "`import numpy as np` when defining a combo_fn."
-                        ),
-                    )
+                second_partial_combo_fn = jit(
+                    grad(lambda x: first_partial_combo_fn(x)[0], holomorphic=True)
+                )
             else:
-                if _optionals.HAS_JAX:
-                    from jax import grad, jit
-
-                    first_partial_combo_fn = jit(grad(operator.combo_fn, holomorphic=True))
-                    second_partial_combo_fn = jit(
-                        grad(lambda x: first_partial_combo_fn(x)[0], holomorphic=True)
-                    )
-                else:
-                    raise MissingOptionalLibraryError(
-                        libname="jax",
-                        name="get_hessian",
-                        msg=(
-                            "This automatic differentiation function is based on JAX. "
-                            "Please install jax and use `import jax.numpy as jnp` instead "
-                            "of `import numpy as np` when defining a combo_fn."
-                        ),
-                    )
+                first_partial_combo_fn = jit(grad(operator.combo_fn, holomorphic=True))
+                second_partial_combo_fn = jit(
+                    grad(lambda x: first_partial_combo_fn(x)[0], holomorphic=True)
+                )
 
             # For a general combo_fn F(g_0, g_1, ..., g_k)
             # dF/d θ0,θ1 = sum_i: (∂F/∂g_i)•(d g_i/ d θ0,θ1) + (∂F/∂^2 g_i)•(d g_i/d θ0)•(d g_i/d