<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add a release note file using reno if the change needs to be
  documented in the release notes.
⚠️ If your pull request fixes an open issue, please link to the issue.

- [ ] I have added the tests to cover my changes.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the CONTRIBUTING document.
-->

### Summary

- Speed up `import qiskit` significantly - 1.17s to 0.57s (2x faster) for a Python environment with almost all the optional dependencies, and 0.78s to 0.52s (33% faster) for a Python environment with _only_ Terra and its requirements.
- Add a new module `qiskit.utils.optionals`, with many lazy-evaluation `HAS_X` variables (e.g. `HAS_MATPLOTLIB`, `HAS_AER`), which can be used for lazy testing of optional dependencies, and unified handling of errors if they are not present.


### Details and comments

These two commits significantly speed up `import qiskit` by deferring many imports to their call sites.  In particular, no component of Scipy is loaded, nor is `pkg_resources` (which is surprisingly heavy).  I didn't change any `scipy` imports in algorithms or opflow, because these aren't imported by default, and I thought they might be a bit more work for lesser benefit.

The first commit in this patch also introduces many new `HAS_X` variables, to unify handling of optional dependencies all the way through the codebase.  Previously, many modules were doing ad-hoc testing with
```python
try:
    import optional_dependency
    HAS_OPTIONAL = True
except ImportError:
    HAS_OPTIONAL = False

class MyClass:
    def __init__(self):
        if not HAS_OPTIONAL:
            raise MissingOptionalLibraryError(...)
```

This form has a couple of downsides:

1. We try to import the optional dependency at import time, which adds to the weight of `import qiskit` if the dependency is installed, even if it won't be used.
2. It clutters many functions with error-handling code, even though they are doing the same thing.
3. Several modules were repeating the exact same code in a couple of places.

The first commit introduces some new lazy-loader objects (`LazyImportTester` and `LazySubprocessTester`), which can be evaluated as Booleans, and will lazily test the presence of the dependencies.  This is similar to the old `HAS_MATPLOTLIB` variable.  There is a new module `qiskit.utils.optionals` that exposes many `HAS_X` variables of these types.  In addition to the lazy Boolean checkers, they also come with unified error handling: modules now can do
```python
@HAS_MATPLOTLIB.require_in_call
def needs_matplotlib(self):
    import matplotlib

def chooses_based_on_argument(self, arg):
    if arg == "text":
        ...
    elif arg == "matplotlib":
        HAS_MATPLOTLIB.require_now("matplotlib drawer")
        import matplotlib
        ...
```

Both `require` forms will raise a suitable `MissingOptionalLibraryError` if the dependency is not available, at the point of call - the decorator raises when the inner function is called, not when the decorator is applied.

Moving the variables into `qiskit.utils.optionals` is also nice for the test suite - instead of it having to repeat the code or import names from random locations, there's a simple namespace where all the objects exist (with documentation on each one in Sphinx).

Fix #7498, as a side-effect - all the instances where initialisation code was logging are now deferred to their usage sites, so the logs will actually make sense and be useful.  You can import every single module defined in Terra without us issuing a log message now.