@woodsp-ibm 2022-01-13
I guess one downside of the new way, from a user perspective, I write code with a bunch of imports and it all seems fine as it loads and starts running. But instead of it failing fast as it used to, I now find it fails some time later if some optional dependency is not present when it finally goes to construct/uses something. The code may have expended significant computational resources by that point. And I might fix that and run again, go further and hit another missing optional. The fail early that was done before would catch the optional dependence early; but has other considerations that you point out such as loading resources that are never used in practice.