@jakelishman 2022-01-13
This doesn't actually change anything about that - before, all the imports would silently fail immediately, setting (e.g.) `_HAS_SKQUANT = False`, and then you wouldn't get an error til you instantiated the class.  In this form, you still get the error in the exact same place, but you also don't pay the import cost til then either.