@woodsp-ibm 2022-01-13
Ah right, as you say the behavior in terms of it failing later is unaltered by your change. I know in Aqua we used to indicate the failure to load but guess that was removed as it may have been too noisy if you did not care about installing the optional.