@jakelishman 2022-01-13
Yeah, I don't know when it changed, but it was probably the right call - this PR identified 26 different optional dependencies across Terra alone.  The code from Aqua seems to be responsible for about 4 of them, which is a bit more reasonable, but still, they're pretty niche optimisers, and I'd imagine most people wouldn't be bothered.  In general, you should get the error when you instantiate the optimiser classes, which is likely going to be one of the first things you do.