@jlapeyre 2022-01-21
A couple of lines in `CONTRIBUTING.md` might be useful. Something about when to add a dependency to  `requirements.txt` and when/how to add it to `optionals.py`.
(This looks so useful, I'm surprised there's no package for it, or at least how-to for something similar)