@jakelishman 2022-01-24
Added a section into `CONTRIBUTING` in fd74223.

The dependency-checker classes themselves I suppose could be split off into a separate package, but there would be some additional work around making the error handling nice and extensible there.  I think the majority of the benefit comes from defining all the `HAS_X` variables, and those would still need to be written by each project, as in `utils.optionals`.